Move "square.cluster_config" to "k8s.cluster_config".
