+++ square/square.py
@@ -386,7 +386,7 @@ def apply_plan(
     """
     try:
         # Create properly configured Requests session to talk to K8s API.
-        (k8s_config, k8s_client), err = cluster_config(kubeconfig, kube_ctx)
+        (k8s_config, k8s_client), err = k8s.cluster_config(kubeconfig, kube_ctx)
         assert not err and k8s_config and k8s_client
 
         # Create the missing resources.
@@ -445,7 +445,7 @@ def make_plan(
     """
     try:
         # Create properly configured Requests session to talk to K8s API.
-        (k8s_config, k8s_client), err = cluster_config(kubeconfig, kube_ctx)
+        (k8s_config, k8s_client), err = k8s.cluster_config(kubeconfig, kube_ctx)
         assert not err and k8s_config and k8s_client
 
         # Load manifests from local files.
@@ -505,7 +505,7 @@ def get_resources(
 
     try:
         # Create properly configured Requests session to talk to K8s API.
-        (k8s_config, k8s_client), err = cluster_config(kubeconfig, kube_ctx)
+        (k8s_config, k8s_client), err = k8s.cluster_config(kubeconfig, kube_ctx)
         assert not err and k8s_config and k8s_client
 
         # Load manifests from local files.
@@ -529,46 +529,3 @@ def get_resources(
 
     # Success.
     return (None, False)
-
-
-def cluster_config(
-        kubeconfig: Filepath,
-        context: Optional[str]) -> Tuple[Tuple[Optional[K8sConfig], Any], bool]:
-    """Return web session to K8s API.
-
-    This will read the Kubernetes credentials, contact Kubernetes to
-    interrogate its version and then return the configuration and web-session.
-
-    Inputs:
-        kubeconfig: str
-            Path to kubeconfig file.
-        kube_context: str
-            Kubernetes context to use (can be `None` to use default).
-
-    Returns:
-        Config, client
-
-    """
-    # Read Kubeconfig file and use it to create a `requests` client session.
-    # That session will have the proper security certificates and headers so
-    # that subsequent calls to K8s need not deal with it anymore.
-    kubeconfig = kubeconfig.expanduser()
-    try:
-        # Parse Kubeconfig file.
-        config = k8s.load_auto_config(kubeconfig, context, disable_warnings=True)
-        assert config
-
-        # Configure web session.
-        client = k8s.session(config)
-        assert client
-
-        # Contact the K8s API to update version field in `config`.
-        config, err = k8s.version(config, client)
-        assert not err and config
-    except AssertionError:
-        return ((None, None), True)
-
-    # Log the K8s API address and version.
-    logit.info(f"Kubernetes server at {config.url}")
-    logit.info(f"Kubernetes version is {config.version}")
-    return (config, client), False
