+++ tests/conftest.py
@@ -12,6 +12,6 @@ def pytest_configure(*args, **kwargs):
 
 @pytest.fixture
 def kube_creds(request):
-    with mock.patch.object(square.square, "cluster_config") as m:
+    with mock.patch.object(square.k8s, "cluster_config") as m:
         m.return_value = (("k8s_config", "k8s_client"), False)
         yield m
